backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
asyncpg
pdfplumber
google-generativeai
python-dotenv
